+++ src/main/java/roomescape/controller/admin/AdminReservationController.java
@@ -1,45 +1,28 @@
-package roomescape.controller;
+package roomescape.controller.admin;
 
 import jakarta.validation.Valid;
 import java.net.URI;
+import java.util.List;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import roomescape.service.ReservationService;
+import roomescape.service.dto.ReservationConditionRequest;
 import roomescape.service.dto.ReservationResponse;
 import roomescape.service.dto.ReservationSaveRequest;
 
 @Controller
 @RequestMapping("/admin")
-public class AdminController {
+public class AdminReservationController {
 
     private final ReservationService reservationService;
 
-    public AdminController(ReservationService reservationService) {
+    public AdminReservationController(ReservationService reservationService) {
         this.reservationService = reservationService;
-    }
-
-    @GetMapping
-    public String home() {
-        return "/admin/index";
-    }
-
-    @GetMapping("/reservation")
-    public String reservation() {
-        return "/admin/reservation-new";
-    }
-
-    @GetMapping("/time")
-    public String time() {
-        return "/admin/time";
-    }
-
-    @GetMapping("/theme")
-    public String theme() {
-        return "/admin/theme";
     }
 
     @PostMapping("/reservations")
@@ -49,5 +32,13 @@ public class AdminController {
         ReservationResponse reservationResponse = reservationService.saveReservation(reservationSaveRequest);
         return ResponseEntity.created(URI.create("/reservations/" + reservationResponse.id()))
                 .body(reservationResponse);
+    }
+
+    @GetMapping("/reservations")
+    public ResponseEntity<List<ReservationResponse>> getReservations(@ModelAttribute @Valid ReservationConditionRequest reservationConditionRequest) {
+        List<ReservationResponse> reservationResponses = reservationService.findReservationsByCondition(reservationConditionRequest);
+
+        return ResponseEntity.ok()
+                .body(reservationResponses);
     }
 }
